broverse-memory/requirements.txt
openai>=1.12.0
chromadb>=0.4.22
click>=8.1.7
pydantic>=2.5.3
python-dotenv>=1.0.0
